Add native libvirt TechVault live evidence
