+++ changelog.d/601.added.md
@@ -4,7 +4,7 @@
 - Tightened the libvirt backend type and reconciliation helpers so SonarCloud accepts the new backend surface.
 - Validated the TechVault scenario through dynamic instantiation, planning, and libvirt provisioning, including switch-backed network links.
 - Added the full TechVault operational scenario and libvirt provisioning coverage for its 30-node, four-network SOC/enterprise/red-team surface.
-- Added `aces libvirt techvault validate-live`, which boots the TechVault operational scenario through the ACES/libvirt provisioning path and verifies container readiness, Kali reachability, Wazuh telemetry, Suricata telemetry, Wazuh agent readiness, SOC readback, and run-archive evidence with focused helper coverage.
+- Added APTL-style reduced TechVault scenario variants and native libvirt coverage proving they realize distinct ACES-derived domain/network/service surfaces.
+- Added `aces libvirt techvault validate-live`, which boots TechVault scenarios as native libvirt/QEMU initramfs appliances, verifies the independent substrate, service readiness, Kali-shared-network reachability, SOC readback, clean-boot recomposition, and run-archive evidence.
 - Added an `aces_operations` package for live operational gates so the CLI can invoke TechVault parity checks without crossing backend/runtime ownership boundaries directly.
 - Hardened the TechVault live gate implementation structure so SonarCloud complexity checks stay green while preserving the SOC readiness and evidence checks.
-- Typed the TechVault APTL helper lifecycle boundary so the live gate remains SonarCloud-clean without changing runtime behavior.
